+++ src/int.rs
@@ -17,6 +17,7 @@ mod bit_not;
 mod bit_or;
 mod bit_xor;
 mod cmp;
+mod ct;
 mod div;
 mod div_unsigned;
 mod encoding;
@@ -28,7 +29,6 @@ mod mul;
 mod mul_unsigned;
 mod neg;
 mod resize;
-mod select;
 mod shl;
 mod shr;
 mod sign;
