+++ src/limb.rs
@@ -8,11 +8,11 @@ mod bit_or;
 mod bit_xor;
 mod bits;
 mod cmp;
+mod ct;
 mod encoding;
 mod from;
 mod mul;
 mod neg;
-mod select;
 mod shl;
 mod shr;
 mod sub;
