+++ src/modular/monty_form/ct.rs
@@ -1,6 +1,36 @@
+//! Constant-time support: impls of `Ct*` traits and constant-time `const fn` operations.
+
 use super::MontyParams;
-use crate::modular::MontyForm;
-use crate::{Choice, CtSelect, Odd, U64, Uint};
+use crate::{Choice, CtEq, CtSelect, Odd, U64, Uint, modular::MontyForm};
+
+impl<const LIMBS: usize> CtEq for MontyForm<LIMBS> {
+    fn ct_eq(&self, other: &Self) -> Choice {
+        self.montgomery_form.ct_eq(&other.montgomery_form) & self.params.ct_eq(&other.params)
+    }
+}
+
+impl<const LIMBS: usize> CtEq for MontyParams<LIMBS> {
+    fn ct_eq(&self, other: &Self) -> Choice {
+        self.modulus.ct_eq(&other.modulus)
+            & self.one.ct_eq(&other.one)
+            & self.r2.ct_eq(&other.r2)
+            & self.mod_inv.ct_eq(&other.mod_inv)
+    }
+}
+
+#[cfg(feature = "subtle")]
+impl<const LIMBS: usize> subtle::ConstantTimeEq for MontyForm<LIMBS> {
+    fn ct_eq(&self, other: &Self) -> subtle::Choice {
+        CtEq::ct_eq(self, other).into()
+    }
+}
+
+#[cfg(feature = "subtle")]
+impl<const LIMBS: usize> subtle::ConstantTimeEq for MontyParams<LIMBS> {
+    fn ct_eq(&self, other: &Self) -> subtle::Choice {
+        CtEq::ct_eq(self, other).into()
+    }
+}
 
 impl<const LIMBS: usize> CtSelect for MontyForm<LIMBS> {
     fn ct_select(&self, other: &Self, choice: Choice) -> Self {
