+++ src/uint.rs
@@ -32,6 +32,7 @@ mod bit_xor;
 mod bits;
 mod cmp;
 mod concat;
+mod ct;
 mod div;
 pub(crate) mod div_limb;
 pub(crate) mod encoding;
@@ -48,7 +49,6 @@ mod neg_mod;
 mod pow;
 mod ref_type;
 mod resize;
-mod select;
 mod shl;
 mod shr;
 mod split;
