+++ src/uint/boxed.rs
@@ -8,6 +8,7 @@ mod bit_or;
 mod bit_xor;
 mod bits;
 mod cmp;
+mod ct;
 pub(crate) mod div;
 pub(crate) mod encoding;
 mod from;
@@ -18,7 +19,6 @@ mod mul_mod;
 mod neg;
 mod neg_mod;
 mod pow_mod;
-mod select;
 mod shl;
 mod shr;
 mod sqrt;
