+++ src/uint/boxed/cmp.rs
@@ -5,7 +5,7 @@
 pub(super) use core::cmp::{Ordering, max};
 
 use super::BoxedUint;
-use crate::{Choice, CtEq, CtGt, CtLt, CtSelect, Limb, Uint, word};
+use crate::{CtEq, CtGt, CtLt, CtSelect, Limb, Uint};
 
 impl BoxedUint {
     /// Returns the Ordering between `self` and `rhs` in variable time.
@@ -33,38 +33,6 @@ impl BoxedUint {
     /// Determine in variable time whether the `self` is zero.
     pub(crate) fn is_zero_vartime(&self) -> bool {
         !self.limbs.iter().any(|l| l.0 != 0)
-    }
-}
-
-impl CtEq for BoxedUint {
-    #[inline]
-    fn ct_eq(&self, other: &Self) -> Choice {
-        let limbs = max(self.nlimbs(), other.nlimbs());
-        let mut ret = Choice::TRUE;
-
-        for i in 0..limbs {
-            let a = self.limbs.get(i).unwrap_or(&Limb::ZERO);
-            let b = other.limbs.get(i).unwrap_or(&Limb::ZERO);
-            ret &= a.ct_eq(b);
-        }
-
-        ret
-    }
-}
-
-impl CtGt for BoxedUint {
-    #[inline]
-    fn ct_gt(&self, other: &Self) -> Choice {
-        let (_, borrow) = other.borrowing_sub(self, Limb::ZERO);
-        word::choice_from_mask(borrow.0)
-    }
-}
-
-impl CtLt for BoxedUint {
-    #[inline]
-    fn ct_lt(&self, other: &Self) -> Choice {
-        let (_, borrow) = self.borrowing_sub(other, Limb::ZERO);
-        word::choice_from_mask(borrow.0)
     }
 }
 
@@ -108,84 +76,10 @@ impl<const LIMBS: usize> PartialOrd<Uint<LIMBS>> for BoxedUint {
     }
 }
 
-#[cfg(feature = "subtle")]
-impl subtle::ConstantTimeEq for BoxedUint {
-    #[inline]
-    fn ct_eq(&self, other: &Self) -> subtle::Choice {
-        CtEq::ct_eq(self, other).into()
-    }
-}
-
-#[cfg(feature = "subtle")]
-impl subtle::ConstantTimeGreater for BoxedUint {
-    #[inline]
-    fn ct_gt(&self, other: &Self) -> subtle::Choice {
-        CtGt::ct_gt(self, other).into()
-    }
-}
-
-#[cfg(feature = "subtle")]
-impl subtle::ConstantTimeLess for BoxedUint {
-    #[inline]
-    fn ct_lt(&self, other: &Self) -> subtle::Choice {
-        CtLt::ct_lt(self, other).into()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::BoxedUint;
-    use crate::{CtEq, CtGt, CtLt};
     use core::cmp::Ordering;
-
-    #[test]
-    fn ct_eq() {
-        let a = BoxedUint::zero();
-        let b = BoxedUint::one();
-
-        assert!(a.ct_eq(&a).to_bool());
-        assert!(!a.ct_eq(&b).to_bool());
-        assert!(!b.ct_eq(&a).to_bool());
-        assert!(b.ct_eq(&b).to_bool());
-    }
-
-    #[test]
-    fn ct_gt() {
-        let a = BoxedUint::zero();
-        let b = BoxedUint::one();
-        let c = BoxedUint::max(64);
-
-        assert!(b.ct_gt(&a).to_bool());
-        assert!(c.ct_gt(&a).to_bool());
-        assert!(c.ct_gt(&b).to_bool());
-
-        assert!(!a.ct_gt(&a).to_bool());
-        assert!(!b.ct_gt(&b).to_bool());
-        assert!(!c.ct_gt(&c).to_bool());
-
-        assert!(!a.ct_gt(&b).to_bool());
-        assert!(!a.ct_gt(&c).to_bool());
-        assert!(!b.ct_gt(&c).to_bool());
-    }
-
-    #[test]
-    fn ct_lt() {
-        let a = BoxedUint::zero();
-        let b = BoxedUint::one();
-        let c = BoxedUint::max(64);
-
-        assert!(a.ct_lt(&b).to_bool());
-        assert!(a.ct_lt(&c).to_bool());
-        assert!(b.ct_lt(&c).to_bool());
-
-        assert!(!a.ct_lt(&a).to_bool());
-        assert!(!b.ct_lt(&b).to_bool());
-        assert!(!c.ct_lt(&c).to_bool());
-
-        assert!(!b.ct_lt(&a).to_bool());
-        assert!(!c.ct_lt(&a).to_bool());
-        assert!(!c.ct_lt(&b).to_bool());
-    }
 
     #[test]
     fn cmp() {
